Add File Reader in C# and Refactor C++ Project
I deleted any files having to do with the preliminary File reader in C++ and I added a whole new project for a windows forms application that reads a file in C#. The file format is xlsx or xls. There is also a sample of the scheduling file in the same folder.